backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
langchain
langchain-google-genai
langgraph
pydantic
python-dotenv
pyyaml
aiosqlite
